requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
click-default-group==1.2.4
cryptography==46.0.3
fastapi==0.119.0
fastapi-login==1.10.3
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
qrcode==8.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.48.0
tabulate==0.9.0
types-Pillow==10.2.0.20240822
types-qrcode==8.2.0.20250914
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
